fix: prevent improper clobbering of man/bin links
This uses the updated version of gentle-fs which has the binLink method,
so that we don't need to fork based on OS, and will always check to
ensure that the bin or cmd-shim is a reference into the folder being
linked.

Also perform a similar check on linked man pages.

Fix #11

PR-URL: https://github.com/npm/bin-links/pull/12
Credit: @isaacs
Close: #12
Reviewed-by: @isaacs